emr/eks/videos/custom_images/requirements.txt
appdirs==1.4.4
black==21.5b2
bokeh==2.3.2
click==8.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mypy-extensions==0.4.3
numpy==1.20.3
packaging==20.9
pathspec==0.8.1
Pillow==8.2.0
py4j==0.10.9
pyparsing==2.4.7
pyspark==3.1.1
python-dateutil==2.8.1
PyYAML==5.4.1
regex==2021.4.4
Shapely==1.7.1
six==1.16.0
toml==0.10.2
tornado==6.1
typing-extensions==3.10.0.0
